// src/app/models/Cattle.ts
export interface Cattle{
    id?: any;
    farmid: any,
    cattleTagId: string;
    cattleBreed: string;
    cattleDOB: string;
    specialFeature: String;
    sex: string;
    noLactation: string;
    birthWeg: string;
    breedingWeg: string;
    cattleWeaningWeg: string;
    avgPreWeg: string;
    avgPostWeg: string; 
    lastCalvingDate: string;
    cattleImg: string;
  }
  